tools/discord_agent/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
pillow>=9.0.0
pytesseract>=0.3.10
